+++ client/src/lib/waiverUtils.ts
@@ -1,82 +1,106 @@
-// place files you want to import through the `$lib` alias in this folder.
-import { getStroke } from 'perfect-freehand-for-svelte'
-import { cubicInOut } from 'svelte/easing'
-import { getSvgPathFromStroke } from './svg-path'
+export const uploadWaiver = async (file, clientId, documentName, layers, width, height, preview) => {
+  try {
+  const clientIdSuffix = clientId.slice(-5);
+  const nameParts = documentName.split(' ');
+  const lastName = nameParts.length > 1 ? nameParts[1] : '';
+  const firstName = nameParts.length > 0 ? nameParts[0] : '';
+  const fileName = `waiver-${lastName}-${firstName}-${clientIdSuffix}.html`;
+  // Constructing the filename for the Blob
+  const blobFileName = `waiver-${lastName}-${firstName}-${clientIdSuffix}.html`;
 
-const linear = (t: number) => t
 
-const strokeOptions = {
-  size: 8,
-  thinning: 0.7,
-  smoothing: 0.4,
-  streamline: 0.6,
-  easing: linear,
-  start: {
-    taper: 30,
-    easing: cubicInOut,
-    cap: true,
-  },
-  end: {
-    taper: 10,
-    easing: cubicInOut,
-    cap: true,
-  },
-}
+  const participantName = (document.getElementById('participantName') as HTMLInputElement)?.value;
+  const address = (document.getElementById('address') as HTMLInputElement)?.value;
+  const guardianName = (document.getElementById('guardianName') as HTMLInputElement)?.value;
+  const date = (document.getElementById('date') as HTMLInputElement)?.value;
+  const signatureLayer = layers.map(layer => layer.path).join('');
 
-interface Options {
-  ondraw: (path: string) => void
-  oncomplete: (path: string) => void
-}
+  const content = `
+    <html>
+      <head>
+        <title>The Pump Factory Waiver</title>
+        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/tailwindcss@2.2.19/dist/tailwind.min.css">
+      </head>
+      <body>
+        <main class="w-5/6 m-auto leading-8">
+          <!-- Hard-coded HTML content with participantName, guardianName, and date -->
+          <h1 class='text-5xl text-center'>The Pump Factory</h1>
+          <h1 class='text-2xl my-8 text-center'>Waiver and Release of Liability</h1>
 
-export function signature(node: HTMLElement, options: Options) {
-  const points: number[][] = []
+<p class="text-xl uppercase leading-8">Please read this waiver and release of liability carefully before participating in any activities at The Pump Factory Rock Climbing Gym.</p>
 
-  function render(complete: boolean) {
-    const stroke = getStroke(points, strokeOptions)
-    const path = getSvgPathFromStroke(stroke)
-    if (complete) {
-      options.oncomplete(path)
+<p class="text-xl uppercase leading-8">By signing this waiver and release of liability, you agree to release and discharge the gym, staff, or any of its owners from any and all liability arising out of your use of the facilities and participation in but not limited to rock climbing activities (i.e., Top Rope Climbing, Lead Climbing, Bouldering, etc.).</p>
+<div>
+  <p>I, <strong>${participantName}</strong>, of legal age, and resident of <strong>${address}</strong>, hereby acknowledge that I/my minor am/is voluntarily participating in activities at The Pump Factory Rock Climbing Gym (hereinafter “the Gym”). In consideration to use the facilities and equipment provided, I agree to the following terms and conditions:</p>
+
+
+  <ol>
+    <li><strong>Assumption of Risks:</strong> I understand and acknowledge that rock climbing, bouldering, and related activities involve inherent risks of injury or death, caused by but not limited to falling, equipment failure, and the actions of other participants beyond my control. I voluntarily assume all risks involved;</li>
+    <li><strong>Release of Liability:</strong> I release, waive, and discharge The Pump Factory, its owners, employees, and affiliates from any and all liability, claims, demands, actions, or causes of action arising out of any loss, damage, or injury, including death, that may be sustained by me and/or my minor while participating in activities at the Gym;</li>
+    <li><strong>Responsibility for Personal Property and Damage to Property:</strong> I am responsible for my and my minor’s personal property while on the premises, including the surroundings of the Gym, and I release the Gym from any liability for loss or damage to personal property. I understand that if I or my minor cause damage to any equipment or property of the gym, I may be liable for the cost of repair or replacement subject to the discretion of the management of the Gym;</li>
+    <li><strong>Fitness to Participate:</strong> I hereby confirm that I/my minor am/is physically and mentally capable of participating in rock climbing activities. I have not been advised against participation by a medical professional and do not have any medical conditions or disabilities that would increase the risk associated with rock climbing. Should I fail to disclose any medical condition that I/my minor possess, I hereby assume all liability;</li>
+    <li><strong>Medical Authorization:</strong> In the event of any injury or illness, I authorize the Gym staff to contact my emergency contact, seek, and consent to emergency medical treatment on my behalf. I am responsible for the safety of my minor;</li>
+    <li><strong>Compliance with Rules and Regulations:</strong> I hereby agree to abide by all the rules and guidelines of the Gym and to follow the instructions of the staff in order to have a safe and enjoyable experience at the Gym.</li>
+  </ol>
+
+  <p class="text-xl uppercase mt-8 leading-8">I have carefully read and hereby waive, release, and assume all risk and voluntarily sign this waiver and release of liability agreement.</p>
+  
+          <p><strong>Printed Name of Participant: </strong>${participantName}</p>
+          <p><strong>Printed Name of Guardian: </strong>${guardianName}</p>
+          <label for="date"> 
+            <strong>Signature Date: </strong>${date}
+          </label>
+
+          <!-- Signature layers -->
+          <div class="relative">
+            <svg viewBox="0 0 ${width || 300} ${height || 150}">
+              <path d="${signatureLayer}" />
+          </div>
+        </main>
+      </body>
+    </html>
+  `;
+
+  // Uploading to pump endpoint
+
+  const formData = new FormData();
+  formData.append('file', new Blob([content], { type: 'text/html' }), blobFileName);
+  formData.append('documentType', 'WAIVER'); // Set the documentType
+  formData.append('documentURL', ''); // This can be left empty for now, as the URL will be obtained after upload
+
+    const response = await fetch('https://uploads.thepumpfactory.net/upload', {
+      // const response = await fetch('http://localhost:3000/upload', {
+      method: 'POST',
+      body: formData,
+    });
+
+    if (response.ok) {
+    const resultUpload = await response.json();
+    // Constructing the filename for the mutation
+    const mutationFileName = `waiver-${lastName}-${firstName}-${clientIdSuffix}.html`;
+
+    return {
+      documentURL: `https://uploads.thepumpfactory.net/uploads/${mutationFileName}`,
+      // documentURL: `http://localhost:3000/uploads/${mutationFileName}`,
+      fileName: mutationFileName,
+    };
     } else {
-      options.ondraw(path)
+      console.error('File upload failed');
+      return null
     }
+  } catch (error) {
+    console.error('Error during file upload:', error);
+
   }
 
-  let down = false
-
-  function pointerDown(e: PointerEvent) {
-    node.setPointerCapture(e.pointerId)
-    points.push([e.offsetX, e.offsetY, e.pressure])
-    render(false)
-    down = true
-  }
-
-  function pointerMove(e: PointerEvent) {
-    if (down && e.isPrimary) {
-      points.push([e.offsetX, e.offsetY, e.pressure])
-      render(false)
-    }
-  }
-
-  function pointerUp(e: PointerEvent) {
-    node.releasePointerCapture(e.pointerId)
-
-    render(true)
-
-    down = false
-    points.length = 0
-  }
-
-  node.addEventListener('pointerdown', pointerDown, { passive: true })
-  node.addEventListener('pointermove', pointerMove, { passive: true })
-  node.addEventListener('pointerup', pointerUp, { passive: true })
-  node.addEventListener('pointercancel', pointerUp, { passive: true })
-
-  return {
-    destroy() {
-      node.removeEventListener('pointerdown', pointerDown)
-      node.removeEventListener('pointermove', pointerMove)
-      node.removeEventListener('pointerup', pointerUp)
-      node.removeEventListener('pointercancel', pointerUp)
-    },
-  }
-}
+  //code block below is for saving html locally
+  // const blob = new Blob([content], { type: 'text/html' });
+  // const url = URL.createObjectURL(blob);
+  
+  // const a = document.createElement('a');
+  // a.href = url;
+  // a.download = 'waiver.html';
+  // document.body.appendChild(a);
+  // a.click();
+  // document.body.removeChild(a);
+};
